UI Tests MBank Web ready

+++ src/MBank/Tests/iOS/ProfileTest.php
@@ -6,57 +6,78 @@ class ProfileTest extends \MBank\Tests\MBankiOSTestCase
 
     public function testUploadPhoto()
     {
-
-        $wallet = $this->createWalletAndLoadDashboard();
-//      $this->getAPIService()->setLoadUserData($wallet);
-        // Add photo in profile
-        $this->byElement('Profile_Button')->click();
-        $this->byName('Add photo')->click();
-        $this->byName('From gallery')->click();
-        $this->waitForElementDisplayedByName('OK');
-        $this->byName('OK')->click();
-        // Load photo
-        $this->waitForElementDisplayedByName('Moments');
-        $this->byName('Moments')->click();
-        // Check photo directory
-        $this->waitForElementDisplayedByXPath('//UIAApplication[1]/UIAWindow[2]/UIACollectionView[1]/UIACollectionCell[1]');
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIACollectionView[1]/UIACollectionCell[1]')->click();
-        $this->waitForElementDisplayedByName('OK');
-        $this->byName('OK')->click();
-        // Check photo was actually uploaded in app
-        $this->waitForElementDisplayedByXPath('//UIAApplication[1]/UIAWindow[2]/UIAButton[2]');
-        // Check that photo was actually uploaded
-        sleep(5);
-        $wallet_data = $this->getAPIService()->getWallet($wallet->phone, $wallet->password);
-        $this->assertTrue(array_key_exists('picture_url', $wallet_data['data']), "Can't find profile image");
-        // Delete wallet
-        $this->getAPIService()->deleteWallet($wallet->phone);
+        if (APP_ENV == 'ios') {
+            $wallet = $this->createWalletAndLoadDashboard();
+//          $this->getAPIService()->setLoadUserData($wallet);
+            // Add photo in profile
+            $this->waitForElementDisplayedByElement('Your_balance_Button');
+            $this->byElement('Profile_Button')->click();
+            $this->byElement('Add_Photo_Button')->click();
+            $this->byElement('Gallery_Button')->click();
+            $this->waitForElementDisplayedByElement('OK_Button');
+            $this->byElement('OK_Button')->click();
+            // Load photo
+            $this->waitForElementDisplayedByElement('Moments_Button');
+            $this->byElement('Moments_Button')->click();
+            // Check photo directory
+            $this->waitForElementDisplayedByElement('Photo_Assert');
+            $this->byElement('Photo_Assert')->click();
+            $this->waitForElementDisplayedByElement('OK_Button');
+            $this->byName('OK_Button')->click();
+            // Check photo was actually uploaded in app
+            $this->waitForElementDisplayedByElement('Photo_Load');
+            // Check that photo was actually uploaded
+            sleep(5);
+            $wallet_data = $this->getAPIService()->getWallet($wallet->phone, $wallet->password);
+            $this->assertTrue(array_key_exists('picture_url', $wallet_data['data']), "Can't find profile image");
+            // Delete wallet
+            $this->getAPIService()->deleteWallet($wallet->phone);
+        } elseif (APP_ENV == 'web') {
+            //TODO for WEB_APP
+            $this->markTestSkipped("Issue not resolved for WEB_APP");
+        }
     }
 
     public function testVerify()
     {
-        $wallet = $this->createWalletAndLoadDashboard();
-        $this->byName('Transfer')->click();
-        $this->waitForElementDisplayedByName('Verification');
-        $this->byName('Next')->click();
-        // Set Valid Data
-        $this->fillVerificationForm($wallet);
-        // Check alert messages before personalisation of user data
-        $this->waitForElementDisplayedByName('Thank you! Your information will be reviewed as soon as possible. You will receive a notification after the process will be complete');
-        $this->byName('Back')->click();
-        $this->waitForElementDisplayedByName('Verification');
-        $this->byName('Вернуться')->click();
-        $this->waitForElementDisplayedByName('Profile');
-        // Personified User
-        $this->getAPIService()->verifyWallet($wallet->phone);
-        // Check P2P Button
-        $this->byName('Your balance')->click();
-        $this->byName('Profile')->click();
-        $this->byName('Menu icon')->click();
-        $this->byName('Transfer')->click();
-        $this->waitForElementDisplayedByXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIAButton[2]');
-        // Delete wallet
-        $this->getAPIService()->deleteWallet($wallet->phone);
+        if (APP_ENV == 'ios') {
+            $wallet = $this->createWalletAndLoadDashboard();
+            $this->waitForElementDisplayedByElement('Your_balance_Button');
+            $this->waitForElementDisplayedByElement('Transfer_Button');
+            $this->byElement('Transfer_Button')->click();
+            $this->waitForElementDisplayedByElement('Verification');
+            $this->byElement('Next_Button')->click();
+            // Set Valid Data
+            $this->fillVerificationForm($wallet);
+            // Check alert messages before personalisation of user data
+            $this->waitForElementDisplayedByElement('Alert_Message_RF');
+            $this->byElement('Back_Button')->click();
+            $this->waitForElementDisplayedByElement('Verification');
+            $this->byElement('Back_Button_Rus')->click();
+            $this->waitForElementDisplayedByElement('Profile_Button');
+            // Personified User
+            $this->getAPIService()->verifyWallet($wallet->phone);
+            // Check P2P Button
+            $this->byElement('Your_balance_Button')->click();
+            $this->byElement('Profile_Button')->click();
+            $this->byElement('Menu_Button')->click();
+            $this->byElement('Transfer_Button')->click();
+            $this->waitForElementDisplayedByElement('Assert_Element');
+            // Delete wallet
+            $this->getAPIService()->deleteWallet($wallet->phone);
+        } elseif (APP_ENV == 'web') {
+            $wallet = $this->createWalletAndLoadDashboard();
+            $this->waitForElementDisplayedByElement('Your_balance_Button');
+            $this->waitForElementDisplayedByElement('Transfer_Button');
+            sleep(1);
+            $this->tap(1, 214, 218, 10); //Profile Button
+            $this->waitForElementDisplayedByElement('Cards_Button');
+            $this->waitForElementDisplayedByElement('Verification_Button');
+            $this->byElement('Verification_Button')->click();
+            $this->fillVerificationForm($wallet);
+            //TODO for WEB_APP
+            $this->markTestSkipped("Issue not resolved for WEB_APP");
+        }
     }
 
     /**
@@ -64,23 +85,29 @@ class ProfileTest extends \MBank\Tests\MBankiOSTestCase
      */
     public function testVerifyWithIncorrectData()
     {
-        $wallet = $this->createWalletAndLoadDashboard();
-        $this->byName('Transfer')->click();
-        $this->waitForElementDisplayedByName('Verification');
-        $this->byName('Next')->click();
-        // Set Invalid Data
-        $this->waitForElementDisplayedByXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[1]');
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[1]')->value('lol');
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[2]')->value('name');
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[3]')->value('test');
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[4]')->value('furman');
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[5]')
-             ->value($wallet->person->passport_issued_at);
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[6]')->value("44/444");
-        $this->byName('Next')->click();
-        $this->waitForElementDisplayedByName('Invalid personal number');
-        // Delete wallet
-        $this->getAPIService()->deleteWallet($wallet->phone);
+        if (APP_ENV == 'ios') {
+            $wallet = $this->createWalletAndLoadDashboard();
+            $this->waitForElementDisplayedByElement('Your_balance_Button');
+            $this->byElement('Transfer_Button')->click();
+            $this->waitForElementDisplayedByElement('Verification_Button');
+            $this->byElement('Next')->click();
+            // Set Invalid Data
+            $this->waitForElementDisplayedByElement('Family_name');
+            $this->byElement('Family_name')->value('lol');
+            $this->byElement('Given_name')->value('name');
+            $this->byElement('Patronymic_name')->value('test');
+            $this->byElement('Passport_series_number')->value('furman');
+            $this->byElement('Passport_issued_at')
+                ->value($wallet->person->passport_issued_at);
+            $this->byElement('Itn')->value("44/444");
+            $this->byElement('Next')->click();
+            $this->waitForElementDisplayedByElement('Invalid_Personal_Number_Alert');
+            // Delete wallet
+            $this->getAPIService()->deleteWallet($wallet->phone);
+        } elseif (APP_ENV == 'web') {
+            //TODO for WEB_APP
+            $this->markTestSkipped("Issue not resolved for WEB_APP");
+        }
     }
 
     /**
@@ -88,14 +115,26 @@ class ProfileTest extends \MBank\Tests\MBankiOSTestCase
      */
     public function testLimits()
     {
-        $wallet = $this->createWalletAndLoadDashboard();
-
-        $this->byName('Profile')->click();
-        // Check the limits is displayed
-        $this->byName('View limits')->click();
-        $this->waitForElementDisplayedByXPath('//UIAApplication[1]/UIAWindow[2]/UIATableView[1]/UIATableGroup[1]');
-        // Delete wallet
-        $this->getAPIService()->deleteWallet($wallet->phone);
+        if (APP_ENV == 'ios') {
+            $wallet = $this->createWalletAndLoadDashboard();
+            $this->waitForElementDisplayedByElement('Your_balance_Button');
+            $this->byElement('Profile_Button')->click();
+            // Check the limits is displayed
+            $this->byElement('View_limits')->click();
+            $this->waitForElementDisplayedByElement('Limits_table');
+            // Delete wallet
+            $this->getAPIService()->deleteWallet($wallet->phone);
+        } elseif (APP_ENV == 'web') {
+            $wallet = $this->createWalletAndLoadDashboard();
+            $this->waitForElementDisplayedByElement('Your_balance_Button');
+            sleep(1);
+            $this->tap(1, 214, 218, 10); //Profile Button
+            // Check the limits is displayed
+            $this->byElement('View_limits')->click();
+            $this->waitForElementDisplayedByElement('Limits_table');
+            // Delete wallet
+            $this->getAPIService()->deleteWallet($wallet->phone);
+        }
     }
 
     /**
@@ -103,21 +142,14 @@ class ProfileTest extends \MBank\Tests\MBankiOSTestCase
      */
     protected function fillVerificationForm($wallet)
     {
-        $this->waitForElementDisplayedByXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[1]');
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[1]')
-             ->value($wallet->person->family_name);
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[2]')
-             ->value($wallet->person->given_name);
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[3]')
-             ->value($wallet->person->patronymic_name);
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[4]')
-             ->value($wallet->person->passport_series_number);
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[6]')
-             ->click();
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[5]')
-             ->value($wallet->person->passport_issued_at);
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[4]/UIATextField[6]')
-             ->value($wallet->person->itn);
-        $this->byName('Next')->click();
+        $this->waitForElementDisplayedByElement('Family_name');
+        $this->byElement('Family_name')->value($wallet->person->family_name);
+        $this->byElement('Given_name')->value($wallet->person->given_name);
+        $this->byElement('Patronymic_name')->value($wallet->person->patronymic_name);
+        $this->byElement('Passport_series_number')->value($wallet->person->passport_series_number);
+        $this->byElement('Itn')->click();
+        $this->byElement('Passport_issued_at')->value($wallet->person->passport_issued_at);
+        $this->byElement('Itn')->value($wallet->person->itn);
+        $this->byElement('Next_Button')->click();
     }
 }
